+++ lib/libesp32/TTGO_TWatch_Library/src/bma.cpp
@@ -187,7 +187,7 @@ void BMA::attachInterrupt()
 
 bool BMA::set_remap_axes(struct bma423_axes_remap *remap_data)
 {
-    bma423_set_remap_axes(remap_data, &_dev);
+    return bma423_set_remap_axes(remap_data, &_dev);
 }
 
 bool BMA::readInterrupt()