+++ CHANGELOG.md
@@ -11,6 +11,11 @@ All notable changes to this project will be documented in this file.
 - Commands ``TuyaRGB``, ``TuyaEnum`` and ``TuyaEnumList`` (#9769)
 - Zigbee command ``ZbInfo`` and prepare support for EEPROM
 - Support for AS608 optical and R503 capacitive fingerprint sensor
+- Command ``SetOption115 1`` to enable ESP32 MiBle
+- Zigbee command ``ZbLeave`` to unpair a device
+- Command ``SetOption116 1`` to disable auto-query of zigbee light devices (avoids network storms with large groups)
+- Support for Shelly Dimmer 1 and 2 by James Turton (#9854)
+- IRremoteESP8266 library from v2.7.11 to v2.7.12
 
 ### Changed
 - Core library from v2.7.4.5 to v2.7.4.7
@@ -22,6 +27,7 @@ All notable changes to this project will be documented in this file.
 - NTP fallback server functionality (#9739)
 - Telegram group chatid not supported (#9831)
 - KNX buttons, switches and sensors detection regression from v9.1.0 (#9811)
+- GUI MqttUser and MqttPassword updates when TLS is compiled in (#9825)
 
 ### Removed
 - Version compatibility check